requirements.txt
addict==2.4.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
earthengine-api==1.5.22
filelock==3.18.0
fiona==1.10.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
geopandas>=0.12.2
httplib2==0.22.0
idna==3.10
igraph==0.11.9
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.4
rsa==4.9.1
rtree==1.4.0
ruff==0.12.2
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentry-sdk==2.32.0
setproctitle==1.3.6
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tcod==19.0.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.6.11
torch==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
